requirements.txt
Telethon==1.10.1
python-gitlab==1.11.0
